Move non-testing functions into utils
